+++ calendar/calendar-common/tests/rsvp/fetch.rs
@@ -6,8 +6,14 @@ use indoc::indoc;
 use jiff::{Zoned, civil::Weekday};
 use pretty_assertions as pa;
 use proton_calendar_api::ProtonCalendarMock;
-use proton_calendar_common::{RsvpError, RsvpEventId, RsvpIntent, RsvpProgress, RsvpRecency};
-use proton_core_api::session::{Config, Session};
+use proton_calendar_common::{
+    RsvpError, RsvpEventId, RsvpFetchError, RsvpIntent, RsvpProgress, RsvpRecency,
+};
+use proton_core_api::{
+    session::{Config, Session},
+    status_observer::StatusObserver,
+    status_watcher::StatusWatcher,
+};
 use proton_core_common::test_utils::test_context::MockApiEnv;
 use std::str::FromStr;
 
@@ -379,7 +385,12 @@ async fn offline() {
         let env = MockApiEnv::new("http://localhost:1");
         let cfg = Config::for_env(env);
 
-        Session::builder().with_config(&cfg).build().await.unwrap()
+        Session::builder()
+            .with_config(&cfg)
+            .with_status(StatusWatcher::with_observer(StatusObserver::test()))
+            .build()
+            .await
+            .unwrap()
     };
 
     let actual = RsvpEventId::invite(INVITE)
@@ -472,7 +483,10 @@ async fn party_crasher() {
         .await
         .unwrap_err();
 
-    assert!(matches!(actual, RsvpError::NotInvited));
+    assert!(matches!(
+        actual,
+        RsvpFetchError::Rsvp(RsvpError::NotInvited)
+    ));
 }
 
 #[tokio::test]
@@ -519,7 +533,10 @@ async fn err_unknown_attendee() {
         .unwrap_err();
 
     // Attendee `zar@pm.me` is not present in the `CalendarEvent`
-    assert_eq!(RsvpError::UnknownAttendee.to_string(), actual.to_string());
+    assert!(matches!(
+        actual,
+        RsvpFetchError::Rsvp(RsvpError::UnknownAttendee),
+    ));
 }
 
 #[tokio::test]
@@ -563,10 +580,10 @@ async fn err_missing_x_pm_token() {
         .await
         .unwrap_err();
 
-    assert_eq!(
-        RsvpError::AttendeeHasNoXPmToken.to_string(),
-        actual.to_string()
-    );
+    assert!(matches!(
+        actual,
+        RsvpFetchError::Rsvp(RsvpError::AttendeeHasNoXPmToken),
+    ));
 }
 
 #[tokio::test]
@@ -612,8 +629,8 @@ async fn err_many_events_in_ics() {
         .await
         .unwrap_err();
 
-    assert_eq!(
-        RsvpError::IcsContainsMoreThanOneEvent.to_string(),
-        actual.to_string()
-    );
+    assert!(matches!(
+        actual,
+        RsvpFetchError::Rsvp(RsvpError::IcsContainsMoreThanOneEvent),
+    ));
 }