+++ crypto/crypto-calendar/tests/acceptance.rs
@@ -2,59 +2,40 @@ use proton_crypto::crypto::{KeyGeneratorAlgorithm, PGPProviderSync};
 use proton_crypto::{new_pgp_provider, new_srp_provider};
 use proton_crypto_account::keys::{
     KeyFlag, KeyId, LocalAddressKey, LocalUserKey, UnlockedAddressKey, UnlockedAddressKeys,
-    UnlockedUserKey, UnlockedUserKeys,
 };
 use proton_crypto_account::salts::KeySalt;
-use proton_crypto_calendar::{
-    CalendarEventDecryptor, CalendarEventEncryptor, UnlockedCalendarKey, UnlockedKeys,
-};
+use proton_crypto_calendar::{CalendarEventDecryptor, CalendarEventEncryptor, UnlockedCalendarKey};
 
 #[test]
-#[allow(clippy::similar_names)]
 fn export_and_import_key() {
     let pgp1 = new_pgp_provider();
     let pgp2 = new_pgp_provider();
 
-    let (user_key1, address_key1) = key(&pgp1);
-    let (user_key2, address_key2) = key(&pgp1);
-    let (user_key3, address_key3) = key(&pgp1);
+    let address_keys = UnlockedAddressKeys(vec![
+        address_key(&pgp1),
+        address_key(&pgp1),
+        address_key(&pgp1),
+    ]);
 
-    let user_keys = UnlockedUserKeys::from(vec![user_key1, user_key2, user_key3]);
-    let address_keys = UnlockedAddressKeys(vec![address_key1, address_key2, address_key3]);
     let calendar_key = UnlockedCalendarKey::generate(&pgp1).unwrap();
 
-    let keys = UnlockedKeys {
-        user_keys,
-        address_keys,
-    };
-
-    // ---
-    // Case 1: Using user key
-
-    for key in keys.user_keys.iter() {
-        calendar_key
-            .export(&pgp1, key)
-            .unwrap()
-            .import(&pgp2, &keys)
-            .unwrap();
-    }
-
-    // ---
-    // Case 2: Using address key
-
-    for key in keys.address_keys.iter() {
-        calendar_key
-            .export(&pgp1, key)
-            .unwrap()
-            .import(&pgp2, &keys)
-            .unwrap();
-    }
+    calendar_key
+        .export(&pgp1, &address_keys[2])
+        .unwrap()
+        .import(&pgp2, &address_keys)
+        .unwrap();
 }
 
 #[test]
 fn encrypt_and_decrypt_events() {
     let pgp = new_pgp_provider();
-    let address_keys = UnlockedAddressKeys(vec![key(&pgp).1, key(&pgp).1, key(&pgp).1]);
+
+    let address_keys = UnlockedAddressKeys(vec![
+        address_key(&pgp),
+        address_key(&pgp),
+        address_key(&pgp),
+    ]);
+
     let calendar_key = UnlockedCalendarKey::generate(&pgp).unwrap();
 
     // ---
@@ -94,7 +75,7 @@ fn encrypt_and_decrypt_events() {
     assert_eq!(b"Hello, World!", actual.as_slice());
 }
 
-fn key<P>(pgp: &P) -> (UnlockedUserKey<P>, UnlockedAddressKey<P>)
+fn address_key<P>(pgp: &P) -> UnlockedAddressKey<P>
 where
     P: PGPProviderSync,
 {
@@ -110,7 +91,7 @@ where
         .unlock_and_assign_key_id(pgp, KeyId(String::default()), &key_secret)
         .unwrap();
 
-    let address_key = LocalAddressKey::generate(
+    LocalAddressKey::generate(
         pgp,
         "someone@localhost",
         KeyGeneratorAlgorithm::default(),
@@ -120,7 +101,5 @@ where
     )
     .unwrap()
     .unlock_and_assign_key_id(pgp, KeyId(String::new()), &user_key)
-    .unwrap();
-
-    (user_key, address_key)
+    .unwrap()
 }